ppvm_tableau

Module prelude

Re-exports

pub use crate::data::GeneralizedTableau;

pub use crate::data::Tableau;

pub use crate::sparsevec::SparseVector;

pub use crate::tableau_index::TableauIndex;

pub use crate::tableau_like::TableauLike;

Modules

config

Macros

impl_op_mul_assign_coefficient

Implement multiplication by a scalar coefficient for PauliSum. Use this macro to implement MulAssign for other scalar types if needed. This macro will forward the multiplication to the underlying map’s mul_assign method.

Structs

LossyPauliWord

NoStrategy

PauliPattern

A Pauli pattern representing a set of matching Pauli words.

PauliSum

PauliWord

PhasedPauliWord

Enums

Pauli

Represents a Pauli operator (I, X, Y, Z)

Traits

ACMap

ACMapAddAssign

ACMapBase

ACMapConsume

ACMapContains

ACMapInsert

ACMapIter

ACMapMulAssign

ACMapRetain

ACMapScale

AmplitudeDamping

CRx

Clifford

CliffordExtensions

Coefficient

ComplexCoefficient

Config

CorrelatedLossChannel

Depolarizing

Depolarizing2

LossChannel

LossyMeasure

Measure

PauliError

PauliErrorAll

PauliIter

PauliStorage

PauliWordTrait

Word-level Pauli operations. Types implementing this trait automatically gain crate::traits::Clifford and crate::traits::CliffordExtensions behavior via blanket impls in [crate::traits::clifford], using get_lbit / get_xbit / set_xbit / set_zbit / rehash to transform Pauli words.

Projection

Reset

ResetLossChannel

RotationOne

RotationTwo

Strategy

TGate

Trace

Trait for computing trace(self * RHS) if type implements Trace an implementation of TraceBy is also provided.

TwoQubitPauliError

U3Gate